linux-4.1.10-tizen5.0.tgz from https://opensource.samsung.com/main
4c0720d56e559865ed79effd54d2a7de  linux-4.1.10-tizen5.0.tgz
6ae83b6d0c616b5e9ac25cfa097743c8  19_DTV_KantS_KantSU.zip

           133876262 May 13  2019 linux-4.1.10-tizen5.0.tgz
           535618111 Mar 25 00:49 19_DTV_KantS_KantSU.zip

+++ drivers/mtd/mtdpart.c
@@ -452,7 +452,7 @@ static struct mtd_part *allocate_partition(struct mtd_info *master,
 		slave->offset = cur_offset;
 		if (mtd_mod_by_eb(cur_offset, master) != 0) {
 			/* Round up to next erasesize */
-			slave->offset = (mtd_div_by_eb(cur_offset, master) + 1) * master->erasesize;
+			slave->offset = (uint64_t)(mtd_div_by_eb(cur_offset, master) + 1) * master->erasesize;
 			printk(KERN_NOTICE "Moving partition %d: "
 			       "0x%012llx -> 0x%012llx\n", partno,
 			       (unsigned long long)cur_offset, (unsigned long long)slave->offset);
